Refactor path re-maps prefix
* refactor path re-maps prefix `@` to `#`
  * compatibility node native path re-maps
* enhance lint rule: more strong lint rule
