+++ src/__tests__/example.test.ts
@@ -1,4 +1,4 @@
-import { JinEitherFrame } from '@frames/JinEitherFrame';
+import { JinEitherFrame } from '#frames/JinEitherFrame';
 import axios, { type AxiosRequestConfig } from 'axios';
 import type { JinConstructorType } from '../tools/ConstructorType';
