+++ src/frames/JinCreateError.ts
@@ -1,6 +1,6 @@
-import type { JinEitherFrame } from '@frames/JinEitherFrame';
-import type { JinFrame } from '@frames/JinFrame';
-import type { IDebugInfo } from '@interfaces/IDebugInfo';
+import type { JinEitherFrame } from '#frames/JinEitherFrame';
+import type { JinFrame } from '#frames/JinFrame';
+import type { IDebugInfo } from '#interfaces/IDebugInfo';
 import type { AxiosError } from 'axios';
 import httpStatusCodes, { getReasonPhrase } from 'http-status-codes';
