+++ src/interfaces/IHeaderField.ts
@@ -1,4 +1,4 @@
-import type { IHeaderFieldOption } from '@interfaces/IHeaderFieldOption';
+import type { IHeaderFieldOption } from '#interfaces/IHeaderFieldOption';
 
 export interface IHeaderField {
   key: string;
