+++ src/interfaces/IParamFieldOption.ts
@@ -1,5 +1,5 @@
-import type { ICommonFieldOption } from '@interfaces/ICommonFieldOption';
-import type { IQueryParamCommonFieldOption } from '@interfaces/IQueryParamCommonFieldOption';
+import type { ICommonFieldOption } from '#interfaces/ICommonFieldOption';
+import type { IQueryParamCommonFieldOption } from '#interfaces/IQueryParamCommonFieldOption';
 
 export interface IParamFieldOption extends ICommonFieldOption, IQueryParamCommonFieldOption {
   type: 'param';
