+++ src/interfaces/IQueryFieldOption.ts
@@ -1,5 +1,5 @@
-import type { ICommonFieldOption } from '@interfaces/ICommonFieldOption';
-import type { IQueryParamCommonFieldOption } from '@interfaces/IQueryParamCommonFieldOption';
+import type { ICommonFieldOption } from '#interfaces/ICommonFieldOption';
+import type { IQueryParamCommonFieldOption } from '#interfaces/IQueryParamCommonFieldOption';
 
 export interface IQueryFieldOption extends ICommonFieldOption, IQueryParamCommonFieldOption {
   type: 'query';
