+++ src/interfaces/TPassJinEitherFrame.ts
@@ -1,5 +1,5 @@
-import type { JinEitherFrame } from '@frames/JinEitherFrame';
-import type { IDebugInfo } from '@interfaces/IDebugInfo';
+import type { JinEitherFrame } from '#frames/JinEitherFrame';
+import type { IDebugInfo } from '#interfaces/IDebugInfo';
 import type { AxiosResponse } from 'axios';
 import type { IPass } from 'my-only-either';
