+++ src/interfaces/TRequestPart.ts
@@ -1,7 +1,7 @@
-import type { IBodyFieldOption } from '@interfaces/body/IBodyFieldOption';
-import type { IHeaderFieldOption } from '@interfaces/IHeaderFieldOption';
-import type { IParamFieldOption } from '@interfaces/IParamFieldOption';
-import type { IQueryFieldOption } from '@interfaces/IQueryFieldOption';
+import type { IBodyFieldOption } from '#interfaces/body/IBodyFieldOption';
+import type { IHeaderFieldOption } from '#interfaces/IHeaderFieldOption';
+import type { IParamFieldOption } from '#interfaces/IParamFieldOption';
+import type { IQueryFieldOption } from '#interfaces/IQueryFieldOption';
 
 export type TRequestPart =
   | IQueryFieldOption['type']
