+++ src/processors/__tests__/header.param.test.ts
@@ -1,4 +1,4 @@
-import { getHeaderInfo } from '@processors/getHeaderInfo';
+import { getHeaderInfo } from '#processors/getHeaderInfo';
 import { format, parse } from 'date-fns';
 import 'jest';
