+++ src/processors/processBodyFormatters.ts
@@ -1,6 +1,6 @@
-import type { IBodyFieldOption, TMultipleBodyFormatter, TSingleBodyFormatter } from '@interfaces/body/IBodyFieldOption';
-import { applyFormatters } from '@tools/applyFormatters';
-import { isValidPrimitiveType, typeAssert } from '@tools/typeAssert';
+import type { IBodyFieldOption, TMultipleBodyFormatter, TSingleBodyFormatter } from '#interfaces/body/IBodyFieldOption';
+import { applyFormatters } from '#tools/applyFormatters';
+import { isValidPrimitiveType, typeAssert } from '#tools/typeAssert';
 import { get, set } from 'dot-prop';
 import { recursive } from 'merge';
 import type { SetOptional, SetRequired } from 'type-fest';
