+++ src/tools/ConstructorType.ts
@@ -4,7 +4,7 @@
  * [TypeScript: Create a condition-based subset types - DailyJS](https://medium.com/dailyjs/typescript-create-a-condition-based-subset-types-9d902cea5b8c)
  */
 
-import type { AbstractJinFrame } from '@frames/AbstractJinFrame';
+import type { AbstractJinFrame } from '#frames/AbstractJinFrame';
 
 // 1 Transform the type to flag all the undesired keys as 'never'
 type FlagExcludedType<Base, Type> = { [Key in keyof Base]: Base[Key] extends Type ? never : Key };
