+++ src/tools/__tests__/format.test.ts
@@ -1,5 +1,5 @@
 /* eslint-disable max-classes-per-file, no-console, import/no-extraneous-dependencies */
-import { applyFormatters } from '@tools/applyFormatters';
+import { applyFormatters } from '#tools/applyFormatters';
 import { format, parse } from 'date-fns';
 import 'jest';
