+++ tsconfig.json
@@ -17,7 +17,24 @@
     "removeComments": false, 
     "importHelpers": false,                 
     "isolatedModules": true,               
-    "strict": true, 
+    
+    /* Strict Type-Checking Options */
+    "strict": true /* Enable all strict type-checking options. */,
+    "noImplicitAny": true /* Raise error on expressions and declarations with an implied 'any' type. */,
+    "strictNullChecks": true /* Enable strict null checks. */,
+    "strictFunctionTypes": true /* Enable strict checking of function types. */,
+    "strictBindCallApply": true /* Enable strict 'bind', 'call', and 'apply' methods on functions. */,
+    "strictPropertyInitialization": true /* Enable strict checking of property initialization in classes. */,
+    "noImplicitThis": true /* Raise error on 'this' expressions with an implied 'any' type. */,
+    
+    /* Additional Checks */
+    "noUnusedLocals": true /* Report errors on unused locals. */,
+    "noUnusedParameters": true /* Report errors on unused parameters. */,
+    "noImplicitReturns": true /* Report error when not all code paths in function return a value. */,
+    "noFallthroughCasesInSwitch": true /* Report errors for fallthrough cases in switch statement. */,
+    "noUncheckedIndexedAccess": true,
+    
+    /* Module Resolution Options */
     "moduleResolution": "node", 
     "esModuleInterop": true, 
     "experimentalDecorators": true, 
@@ -25,17 +42,12 @@
     "outDir": "dist",
     "baseUrl": ".",
     "paths": {
-      "@tools/*": ["src/tools/*"],
-      "@processors/*": ["src/processors/*"],
-      "@frames/*": ["src/frames/*"],
-      "@interfaces/*": ["src/interfaces/*"]
+      "#tools/*": ["src/tools/*"],
+      "#processors/*": ["src/processors/*"],
+      "#frames/*": ["src/frames/*"],
+      "#interfaces/*": ["src/interfaces/*"]
     }
   },
-  "include": [
-    "**/*.ts"
-  ],
-  "exclude": [
-    "**/dist/**",
-    ".configs/**"
-  ]
+  "include": ["src/**/*.ts"],
+  "exclude": ["examples/**/*", "dist/**", "**/.configs/**", "**/docs/**/*"]
 }
