+++ tsconfig.prod.json
@@ -1,16 +1,18 @@
 {
   "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "rootDir": "src",
-  },
   "exclude": [
-    "node_modules",
+    "src/testenv/*",
+    "**/node_modules/**",
+    "examples/**/*",
     "dist/**",
-    "examples/**",
-    "jest.config.js",
-    ".eslintrc.js",
-    "**/__test__/*",
-    "**/__tests__/*",
-    "**/.configs/**/*"
+    "**/.configs/**",
+    "**/docs/**/*",
+    "**/__test__/**",
+    "**/__tests__/**",
+    "dist/**",
+    "**/.configs/**",
+    "**/docs/**",
+    "jest.config.cjs",
+    "prepublish.cjs"
   ]
 }
